concat check-order and enforce-order using command line argument for enforce
