+++ check-order.py
@@ -2,7 +2,18 @@ import frontmatter
 from pathlib import Path, PosixPath
 import sys
 
+def enforceOrder(listToBeOrdered):
+    for index, post in enumerate(listToBeOrdered):
+        if index+1 != post['order']:      
+            postToBeAltered = frontmatter.load(post['path'])
+            postToBeAltered.metadata['jupyter']['plotly']['order'] = index+1
+            frontmatter.dump(postToBeAltered, post['path'])
+
+def checkConsecutive(l): 
+        return sorted(l) == list(range(1, len(l)+1)) 
+
 categories = ["file_settings", "basic", "statistical", "scientific", "maps", "3d_charts", "multiple_axes"]
+
 try:
     path = str(sys.argv[1])
 except:
@@ -18,31 +29,37 @@ for category in categories:
                 if post.metadata['jupyter']['plotly']['display_as'] == category:
                     postFamily.append({'path':str(md_path), 'order' : post.metadata['jupyter']['plotly']['order']})
                 
+    postFamily.append({'path':'placeholder', 'order' : 5})
+
     sortedPostFamily = sorted(postFamily, key = lambda i: i['order'])
 
-    order = []
-
-    for post in sortedPostFamily:
-        if post['order'] == 5:
-            raise Exception("Order Check Failed! Post {} cannot have order 5!".format(post['path']))
-        order.append(post['order'])
-
-    print(order)
+    order = [ p['order'] for p in sortedPostFamily ]
 
     if order[0] != 1:
-        raise Exception("Order Check Failed! First post in {} display_as does not have order 1!".format(category))
+        if len(sys.argv) == 3:
+            if sys.argv[2] == 'enforce':
+                print('Order Check Did Not Pass! ENFORCING CORRECT ORDER for {}'.format(category))
+                enforceOrder(sortedPostFamily)
+        else:
+            raise Exception("Order Check Failed! First post in {} display_as does not have order 1!".format(category))
 
-    def checkConsecutive(l): 
-        return sorted(l) == list(range(min(l), max(l)+1)) 
+    
 
-    try:
-        checkConsecutive(order)
-    except:
-        raise Exception("Order Check Failed! Orders in {} display_as are not consecutive integers!!".format(category))
+    if not checkConsecutive(order):
+        if len(sys.argv) == 3:
+            if sys.argv[2] == 'enforce':
+                print('Order Check Did Not Pass! ENFORCING CORRECT ORDER for {}'.format(category))
+                enforceOrder(sortedPostFamily)
+        else: 
+            raise Exception("Order Check Failed! Order front-matter in '{}' display_as are not consecutive integers!!".format(category))
 
     if len(order) != len(set(order)):
-        raise Exception("Order Check Failed! {} display_as has duplicate order frontmatter!!".format(category))
-
+        if len(sys.argv) == 3:
+            if sys.argv[2] == 'enforce':
+                print('Order Check Did Not Pass! ENFORCING CORRECT ORDER for {}'.format(category))
+                enforceOrder(sortedPostFamily)
+        else:         
+            raise Exception("Order Check Failed! {} display_as has duplicate order frontmatter!!".format(category))
 
     print("Order Checks Passed for {} display_as!".format(category))
     order = []
